Add menu story that uses parentRef.
