+++ packages/menu/menu.stories.js
@@ -1,173 +1,6 @@
-import React, { useState } from "react"
-import { Button } from "@rent_avail/controls"
-import { Container, Flex, Box } from "@rent_avail/core"
-import {
-  ConfirmationDialog,
-  Dialog,
-  DialogHeader,
-  DialogTarget,
-} from "@rent_avail/dialog/src"
-import { Menu, MenuTarget, MenuList, MenuItem } from "./src"
+import BasicUsage from "./stories/basic"
+import PortalUsage from "./stories/portal"
+import ParentRefUsage from "./stories/parent-ref"
 
 export default { title: "Packages/Menu" }
-
-export function BasicUsage() {
-  return (
-    <>
-      <Flex sx={{ justifyContent: "space-between" }}>
-        <Menu>
-          <MenuTarget>
-            <Button>open `b l`</Button>
-          </MenuTarget>
-          <MenuList sx={{ zIndex: "0" }}>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-          </MenuList>
-        </Menu>
-        <Menu>
-          <MenuTarget>
-            <Button>open `b r`</Button>
-          </MenuTarget>
-          <MenuList>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-          </MenuList>
-        </Menu>
-      </Flex>
-      <Box sx={{ height: "calc(100vh - 10rem)" }} />
-      <Flex sx={{ justifyContent: "space-between" }}>
-        <Menu>
-          <MenuTarget>
-            <Button>open `t l`</Button>
-          </MenuTarget>
-          <MenuList>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-          </MenuList>
-        </Menu>
-        <Menu>
-          <MenuTarget>
-            <Button>open `t r`</Button>
-          </MenuTarget>
-          <MenuList>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-            <MenuItem>
-              <a href="#">thing</a>
-            </MenuItem>
-          </MenuList>
-        </Menu>
-      </Flex>
-    </>
-  )
-}
-
-export function PortalUsage() {
-  const [open, setOpen] = useState(true)
-  function handleClick() {
-    setOpen((o) => !o)
-  }
-  return (
-    <Container>
-      <Dialog open={open} toggle={handleClick} id="confirmation-id">
-        <DialogTarget>
-          <Button onClick={handleClick}>open dialog</Button>
-        </DialogTarget>
-        <ConfirmationDialog
-          style={{ overflowY: "auto", maxWidth: "75rem", width: "75rem" }}
-        >
-          <Container>
-            <DialogHeader title="Open and scroll" />
-            <Flex sx={{ justifyContent: "space-between" }}>
-              <Menu>
-                <MenuTarget>
-                  <Button>open `b l`</Button>
-                </MenuTarget>
-                <MenuList>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                </MenuList>
-              </Menu>
-              <Menu>
-                <MenuTarget>
-                  <Button>open `b r`</Button>
-                </MenuTarget>
-                <MenuList>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                </MenuList>
-              </Menu>
-            </Flex>
-            <Box sx={{ height: "30rem" }} />
-            <Flex sx={{ justifyContent: "space-between" }}>
-              <Menu>
-                <MenuTarget>
-                  <Button>open `t l`</Button>
-                </MenuTarget>
-                <MenuList>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                </MenuList>
-              </Menu>
-              <Menu>
-                <MenuTarget>
-                  <Button>open `t r`</Button>
-                </MenuTarget>
-                <MenuList>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                  <MenuItem>thing</MenuItem>
-                </MenuList>
-              </Menu>
-            </Flex>
-          </Container>
-          <Box height="0rem" />
-        </ConfirmationDialog>
-      </Dialog>
-    </Container>
-  )
-}
+export { BasicUsage, PortalUsage, ParentRefUsage }
